requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
huggingface-hub==0.14.1
pinecone-client>=3.0.0
langchain==0.0.225
langchain-community
flask==2.3.2
pypdf==3.9.0
python-dotenv==1.0.0
-e .